move app tests
