+++ test/integration/app/built-in-package-check.test.ts
