+++ test/integration/app/determine-editor-version.test.ts
