+++ test/integration/app/get-latest-version.test.ts
@@ -1,10 +1,10 @@
+import { GetLatestVersionFromRegistryPackument } from "../../../src/app/get-latest-version";
 import { DomainName } from "../../../src/domain/domain-name";
 import { UnityPackument } from "../../../src/domain/packument";
 import { Registry } from "../../../src/domain/registry";
 import { SemanticVersion } from "../../../src/domain/semantic-version";
 import { GetRegistryPackument } from "../../../src/io/packument-io";
-import { GetLatestVersionFromRegistryPackument } from "../../../src/app/get-latest-version";
-import { exampleRegistryUrl } from "../domain/data-registry";
+import { exampleRegistryUrl } from "../../unit/domain/data-registry";
 import { mockFunctionOfType } from "./func.mock";
 
 describe("get latest version from registry packument", () => {
