+++ test/integration/app/put-registry-auth.test.ts
@@ -1,7 +1,7 @@
+import { putRegistryAuthIntoUpmConfig } from "../../../src/app/put-registry-auth";
 import { Base64 } from "../../../src/domain/base64";
 import { type UpmConfigContent } from "../../../src/io/upm-config-io";
-import { putRegistryAuthIntoUpmConfig } from "../../../src/app/put-registry-auth";
-import { exampleRegistryUrl } from "../domain/data-registry";
+import { exampleRegistryUrl } from "../../unit/domain/data-registry";
 
 describe("put registry auth into upm config", () => {
   const someEmail = "user@mail.com";
